+++ test/shopify-cli/commands/create_test.rb
@@ -3,15 +3,22 @@ require 'test_helper'
 module ShopifyCli
   module Commands
     class CreateTest < MiniTest::Test
-      def test_without_arguments_calls_help
-        @context.expects(:puts).with(ShopifyCli::Commands::Create.help)
-        run_cmd('create')
+      def setup
+        super
+        @cmd = ShopifyCli::Commands::Create
+        @cmd.ctx = @context
+        @cmd_name = 'create'
       end
 
-      def test_with_project_calls_project
-        Create::Project.any_instance.expects(:call)
-          .with(['new-app'], 'project')
-        run_cmd('create project new-app')
+      def test_without_arguments_calls_help
+        @context.expects(:puts).with(ShopifyCli::Commands::Create.help)
+        @cmd.call([], @cmd_name)
+      end
+
+      def test_with_create_pp
+        Create::App.any_instance.expects(:call)
+          .with(['new-app'], 'app')
+        @cmd.call(['app', 'new-app'], @cmd_name)
       end
     end
   end
