+++ test/shopify-cli/commands/generate/webhook_test.rb
@@ -6,13 +6,21 @@ module ShopifyCli
       class WebhookTest < MiniTest::Test
         include TestHelpers::FakeUI
 
+        def setup
+          super
+          Helpers::AccessToken.stubs(:read).returns('myaccesstoken')
+          @cmd = ShopifyCli::Commands::Generate
+          @cmd.ctx = @context
+          @cmd_name = 'generate'
+        end
+
         def test_with_existing_param
           ShopifyCli::Tasks::Schema.expects(:call).returns(
             JSON.parse(File.read(File.join(ShopifyCli::ROOT, "test/fixtures/shopify_schema.json")))
           )
           @context.expects(:system).with('a command')
             .returns(mock(success?: true))
-          run_cmd('generate webhook APP_UNINSTALLED')
+          @cmd.call(['webhook', 'APP_UNINSTALLED'], @cmd_name)
         end
 
         def test_with_incorrect_param_expects_ask
@@ -22,17 +30,17 @@ module ShopifyCli
           CLI::UI::Prompt.expects(:ask).returns('APP_UNINSTALLED')
           @context.expects(:system).with('a command')
             .returns(mock(success?: true))
-          run_cmd('generate webhook create_webhook_fake')
+          @cmd.call(['webhook', 'create_webhook_fake'], @cmd_name)
         end
 
         def test_with_selection
           ShopifyCli::Tasks::Schema.expects(:call).returns(
             JSON.parse(File.read(File.join(ShopifyCli::ROOT, "test/fixtures/shopify_schema.json"))),
-          )
+            )
           CLI::UI::Prompt.expects(:ask).returns('PRODUCT_CREATE')
           @context.expects(:system).with('a command')
             .returns(mock(success?: true))
-          run_cmd('generate webhook')
+          @cmd.call(['webhook'], @cmd_name)
         end
       end
     end
