+++ test/shopify-cli/commands/populate_test.rb
@@ -8,11 +8,13 @@ module ShopifyCli
         Helpers::AccessToken.stubs(:read).returns('myaccesstoken')
         Tasks::Schema.stubs(:call)
         Tasks::EnsureEnv.stubs(:call)
+        @cmd = ShopifyCli::Commands::Populate
+        @cmd.ctx = @context
       end
 
       def test_without_arguments_calls_help
         @context.expects(:puts).with(ShopifyCli::Commands::Populate.help)
-        run_cmd('populate')
+        @cmd.call([], 'populate')
       end
     end
   end
