+++ test/shopify-cli/commands/tunnel_test.rb
@@ -3,19 +3,28 @@ require 'test_helper'
 module ShopifyCli
   module Commands
     class TunnelTest < MiniTest::Test
+
+      def setup
+        super
+        Helpers::AccessToken.stubs(:read).returns('myaccesstoken')
+        @cmd = ShopifyCli::Commands::Tunnel
+        @cmd.ctx = @context
+        @cmd_name = 'tunnel'
+      end
+
       def test_auth
         ShopifyCli::Tasks::Tunnel.any_instance.expects(:auth)
-        run_cmd('tunnel auth')
+        @cmd.call(['auth'], @cmd_name)
       end
 
       def test_start
         ShopifyCli::Tasks::Tunnel.any_instance.expects(:call)
-        run_cmd('tunnel start')
+        @cmd.call(['start'], @cmd_name)
       end
 
       def test_stop
         ShopifyCli::Tasks::Tunnel.any_instance.expects(:stop)
-        run_cmd('tunnel stop')
+        @cmd.call(['stop'], @cmd_name)
       end
     end
   end
